% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{gibbs}
\alias{gibbs}
\title{the two-stage Gibbs sampler method for normal distribution}
\usage{
gibbs(n, r)
}
\arguments{
\item{n}{the duplicate of the chain}

\item{r}{the coefficient of two normal distributions}
}
\value{
results of gibbs sampler
}
\description{
a algorithm called Gibbs sampler using in Mento Carlo Markov chain
}
